It's Uno. What can I say?
Uno. Includes number cards/draw+ cards, skip cards, and wild cards. (There is no stacking draw+ cards cuz that's stupid.)

Wild cards currently do not allow the player of the card to choose a color after playing it, and is therefore just a card that the opponent may play any card on. future updates may add the ability to choose the color.